接收必要的事件
